requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
cachetools==4.2.4
class-registry==2.1.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.16
django-crispy-forms==1.14.0
django-filter==22.1
django-gunicorn==0.1.1
django-heroku==0.3.1
django-import-export==2.9.0
django-multi-email-field==0.6.2
django-tables2==2.4.1
django-tinymce==3.5.0
django-unicorn==0.48.0
et-xmlfile==1.1.0
filters==1.3.2
filters-django==1.0.5
gunicorn==20.1.0
import-export==0.2.67.dev6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
orjson==3.8.0
Pillow==9.2.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
